refactor: rename service to provider
