+++ studio/src/app/providers/auth/auth.factory.provider.ts
@@ -2,10 +2,10 @@ import {EnvironmentAppConfig} from '../../types/core/environment-config';
 
 import {EnvironmentConfigService} from '../../services/environment/environment-config.service';
 
-import {AuthProvider} from './auth.service';
-import {AuthOfflineProvider} from './auth.offline.service';
-import {AuthIcProvider} from './auth.ic.service';
-import {AuthFirebaseProvider} from './auth.firebase.service';
+import {AuthProvider} from './auth.provider';
+import {AuthOfflineProvider} from './auth.offline.provider';
+import {AuthIcProvider} from './auth.ic.provider';
+import {AuthFirebaseProvider} from './auth.firebase.provider';
 
 export class AuthFactoryProvider {
   private static instance: AuthProvider;
